+++ packages/util/src/formatDecimals.ts
@@ -56,9 +56,11 @@ export const formatDecimals = (
   // format
 
   return Intl.NumberFormat(locale, {
-    //compact notation (K, M, B) if above 9999
+    // compact notation (K, M, B) if above 9999
     notation: truncatedValue.gt(9999) ? "compact" : "standard",
-    maximumSignificantDigits: digits + (truncatedValue.lt(1) ? 1 : 0),
+    // NOTE: possible values are from `0` to `21`
+    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat/NumberFormat#maximumsignificantdigits
+    maximumSignificantDigits: Math.max(1, Math.min(digits + (truncatedValue.lt(1) ? 1 : 0), 21)),
     ...options,
   }).format(truncatedValue.toNumber())
 }