+++ packages/chaindata-provider/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @talismn/chaindata-provider
 
+## 1.3.1
+
+### Patch Changes
+
+- 7deed17: generate init data
+- 2c395d3: init data
+- 9b5618c: init chain data
+- Updated dependencies [05e1e30]
+  - @talismn/util@0.5.6
+
 ## 1.3.0
 
 ### Minor Changes